kentro

Struct KMeans

pub struct KMeans {
    pub(crate) n_clusters: usize,
    pub(crate) iters: usize,
    pub(crate) euclidean: bool,
    pub(crate) balanced: bool,
    pub(crate) max_balance_diff: usize,
    pub(crate) verbose: bool,
    pub(crate) trained: bool,
    pub(crate) centroids: Option<Array2<f32>>,
    pub(crate) assignments: Vec<usize>,
    pub(crate) cluster_sizes: Array1<f32>,
}

High-performance K-Means clustering implementation

This implementation provides both standard and balanced K-Means clustering with support for Euclidean and cosine similarity metrics.

Examples

use kentro::KMeans;
use ndarray::Array2;

// Create sample data
let data = Array2::from_shape_vec((100, 2), (0..200).map(|x| x as f32).collect()).unwrap();

// Create and train K-Means
let mut kmeans = KMeans::new(3).with_iterations(50).with_verbose(true);
let clusters = kmeans.train(data.view(), None).unwrap();

println!("Found {} clusters", clusters.len());

Fields
n_clusters: usizeiters: usizeeuclidean: boolbalanced: boolmax_balance_diff: usizeverbose: booltrained: boolcentroids: Option<Array2<f32>>assignments: Vec<usize>cluster_sizes: Array1<f32>
Implementations

impl KMeans

pub fn new(n_clusters: usize) -> Self

Create a new K-Means instance

Arguments

n_clusters - The number of clusters (k)

Panics

Panics if n_clusters is 0

pub fn with_iterations(self, iters: usize) -> Self

Set the number of iterations (default: 25)

pub fn with_euclidean(self, euclidean: bool) -> Self

Use Euclidean distance instead of cosine similarity (default: false)

pub fn with_balanced(self, balanced: bool) -> Self

Enable balanced K-Means clustering (default: false)

pub fn with_max_balance_diff(self, max_balance_diff: usize) -> Self

Set maximum balance difference for balanced clustering (default: 16)

pub fn with_verbose(self, verbose: bool) -> Self

Enable verbose output (default: false)

pub fn train( &mut self, data: ArrayView2<'_, f32>, num_threads: Option<usize>, ) -> Result<Vec<Vec<usize>>, KMeansError>

Perform K-Means clustering on the provided data

Arguments

data - The data matrix (n_points × n_dimensions)

num_threads - Number of threads to use (None for automatic)

Returns

A vector of vectors where each inner vector contains the indices of points assigned to the corresponding cluster

pub fn assign( &self, data: ArrayView2<'_, f32>, k: usize, ) -> Result<Vec<Vec<usize>>, KMeansError>

Assign data points to their k nearest clusters

Arguments

data - The data matrix (n_points × n_dimensions)

k - Number of nearest clusters to assign each point to

Returns

A vector of vectors where each inner vector contains the indices of points assigned to the corresponding cluster

pub fn n_clusters(&self) -> usize

Get the number of clusters

pub fn iterations(&self) -> usize

Get the number of iterations

pub fn is_euclidean(&self) -> bool

Check if using Euclidean distance

pub fn is_balanced(&self) -> bool

Check if using balanced clustering

pub fn centroids(&self) -> Option<ArrayView2<'_, f32>>

Get the cluster centroids

Returns None if the model hasn’t been trained yet

pub fn is_trained(&self) -> bool

Check if the model has been trained

impl KMeans

pub(crate) fn sample_rows(&self, data: &ArrayView2<'_, f32>) -> Array2<f32>

pub(crate) fn assign_clusters( &mut self, data: &ArrayView2<'_, f32>, data_norms: Option<&Array1<f32>>, )

pub(crate) fn update_centroids(&mut self, data: &ArrayView2<'_, f32>)

pub(crate) fn postprocess_centroids(&mut self)

pub(crate) fn split_clusters(&mut self, data: &ArrayView2<'_, f32>)

pub(crate) fn compute_cost(&self, data: &ArrayView2<'_, f32>) -> f32

pub(crate) fn balance_clusters( &mut self, data: &ArrayView2<'_, f32>, data_norms: Option<&Array1<f32>>, ) -> Result<(), KMeansError>